+++ CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#1337](https://github.com/alleslabs/celatone-frontend/pull/1337) Support EVM transactions on account details
+
 ### Improvements
 
 - [#1334](https://github.com/alleslabs/celatone-frontend/pull/1334) Upgrade Initia react-app-shell package